Fix logic in mu-loader.php to support plugins with different main file names
Updated mu-loader.php to address an issue where the main plugin file in the mu_plugins directory was required to have the same name as the directory. The refactored code now loops through all PHP files in the directories, uses get_plugin_data to check for a valid plugin header, and includes the files accordingly. This change allows mu-plugins to be added regardless of the main file name.